mm: add generic helper to hint a large folio
To use PTE batching, we want to determine whether the folio mapped by
the PTE is large, thus requiring the use of vm_normal_folio(). We want
to avoid the cost of vm_normal_folio() if the code path doesn't already
require the folio. For arm64, pte_batch_hint() does the job. To generalize
this hint, add a helper which will determine whether two consecutive PTEs
point to consecutive PFNs, in which case there is a high probability that
the underlying folio is large.

[pfalcato: Small fixups]
Signed-off-by: Dev Jain <dev.jain@arm.com>
Signed-off-by: Pedro Falcato <pfalcato@suse.de>